+++ CHANGELOG.md
@@ -18,6 +18,8 @@ All notable, unreleased changes to this project will be documented in this file.
 ### Other changes
 - Add missing descriptions to order module - #14845 by @DevilsAutumn
 
+- Unify how undiscounted prices are handled in orders and checkouts - #14780 by @jakubkuc
+
 # 3.18.0
 
 ### Highlights